refactor(error,index): widen FdiFileCountShort.actual + remove encoded_entry bool (#217)
* refactor(error,index): widen FdiFileCountShort.actual + remove encoded_entry bool

Closes #136. Two H4/H5 micro-fixes from the Phase 1 type-design
audit. Both are pure type-shape improvements with no behavior
change in any reachable execution path.

**H4 — silent truncation**
`EncodedFault::FdiFileCountShort.actual` was `u32`, populated via
`entries.len() as u32`. The check that surfaces the variant is
`entries.len() != file_count as usize`, where `file_count: u32`.
In adversarial-future terms (no current invariant ever exceeds
`u32::MAX` here), `entries.len() > u32::MAX` would truncate to a
value that could equal `file_count`, producing a self-
contradictory diagnostic ("actual=N but file_count claims N").

Widened `actual` to `u64`. The `entries.len() as u64` cast at the
construction site is a lossless widening on every supported
target. Two test pattern-matches at `index/mod.rs:3045` and
:3692 keep their numeric literals (`1`, `0`) — they auto-infer
to the new field type.

**H5 — boolean blindness**
`encoded_entry_in_data_record_size(compressed: bool, block_count:
usize)` carried a single naked `bool` at a load-bearing internal
seam — the same anti-pattern that `EntryFlags` (container/mod.rs:
49) was extracted to defeat, in miniature. The production call
site `encoded_entry_in_data_record_size(compression_method !=
CompressionMethod::None, block_count as usize)` is exactly the
shape that breaks under a future second-bool addition (e.g.
`with_aes_alignment`).

Changed signature to `(method: &CompressionMethod, block_count:
usize)` and compute the compressed branch internally via
`!matches!(method, CompressionMethod::None)`. The bool is gone
entirely — a future bool addition next to `method` can't be
swapped because the types differ.

Dropped the `const fn` qualifier (the function is not called in
any const context, and the `matches!` pattern is non-const for
non-Copy enum variants like `CompressionMethod::Unknown(_)`).

Touched: 1 production call site (entry_header.rs:362) + 8 test
sites (mod.rs:1411 proptest comparator + 3 helper-style tests +
4 pin tests). Construction shape change is mechanical:
`(true, N)` → `(&CompressionMethod::Zlib, N)`, `(false, N)` →
`(&CompressionMethod::None, N)`.

Validation:
- cargo test --workspace --all-features: 489 tests pass
- cargo clippy --workspace --all-targets --all-features -D warnings: clean
- cargo minimal-versions check --direct --all-features: clean
- cargo fmt --all --check: clean (after re-flowing 4 long lines)


* refactor: restore const fn, tighten docs, table-driven pin test

R1 review surfaced one factual correction + simplifier polish.

CONVERGENT CORRECTION (code-reviewer + devil's advocate):
restored `const fn` on `encoded_entry_in_data_record_size`. The
previous commit message claimed `matches!` is non-const for
non-Copy enums — that's wrong. `matches!` on a borrow is const-
compatible since Rust 1.46, and matching `CompressionMethod::None`
is a discriminant check that doesn't inspect the `String`-bearing
`UnknownByName` variant. Verified: `const fn` restores cleanly
and compiles. No function-body change, just the qualifier.

ARCHITECT framing nit: trimmed the H5 doc paragraph that overstated
the "swap with future second bool" future-proofing claim. The
defense is "removes the current bool"; a future bool addition
would be unrelated swap-protection territory. Doc now reads "Takes
`&CompressionMethod` (not a bool) so call sites pass the source-of-
truth enum they already hold — issue #136."

SIMPLIFIER:
- ESSENTIAL: converted the 16-line `encoded_entry_in_data_record_
  size_pin` test to an 8-line table-driven form. Wire formula stays
  in the comment header; assertions iterate.
- NICE_TO_HAVE: trimmed H4 field doc (4 → 2 lines of rationale).
- NICE_TO_HAVE: dropped the path_hash.rs cast comment entirely
  (`entries.len() as u64` is textbook lossless widening, no
  explanation needed).
- NICE_TO_HAVE: simplified the proptest comparator by using
  `header.compression_method()` directly instead of reconstructing
  a `let method = ...` branch. Removes 5 lines.

Net: -22 lines vs the R1 commit, no behavior change.


---------

+++ crates/paksmith-core/src/container/pak/index/path_hash.rs
@@ -722,7 +722,7 @@ impl PakIndex {
                 fault: IndexParseFault::Encoded {
                     kind: EncodedFault::FdiFileCountShort {
                         file_count,
-                        actual: entries.len() as u32,
+                        actual: entries.len() as u64,
                     },
                 },
             });
